Move session token api request to api client
